backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "oseiasdfarias",email = "oseias.dfarias@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic[email] (>=2.12.3,<3.0.0)",
    "fastapi (>=0.120.0,<0.121.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "openai (>=2.6.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
